Implement new pattern system
The patterns can now use thread names and coordinates for matching a
vehicle. The patterns are also now moved over to the config file
instead of being hardcoded into the ASI.

Default patterns have also been updated to include new improved
patterns.

The default config file is only read if the config file is not
present, which means if the "Pattern" key is deleted from the config
file, the patterns will not work.