Overhauled Vehicle Patterns System
The system for deciding random script vehicles has been completely changed, and has been adapted to more closely resemble that which is used in the GTA V project. This more cleanly separates the patterns into its own file, allowing for easier editing and more freedom of vehicle choice. For the user, this will result in a notable increase in variety of script vehicles once the VehiclePatterns file is complete (currently only half done).

Please note that the weapon patterns still need to be implemented, and so all of that code is yet to be updated.

Fixed some log text.

Updated the credits to add additional contributors to the project since the last version.