Implemented random Vigilante player vehicles
As with other vehicle sub-missions, starting Vigilante will now give you a random vehicle. The feature had to be slightly modified to work with Vigilante though (since it's the only vehicle sub-mission allowing you to switch vehicles).

If your original vehicle is destroyed, you still need to enter a police vehicle to proceed, but any police vehicle you enter will also be randomized into something else, so as to maintain the random aspect of the vehicle you do it with.

Any vehicle replacing a police car will be marked as such by the Rainbomizer, so being inside of it will be considered as being inside a police car. Only the last 3 vehicles will be remembered though, so as not to overload the vehicle pool (plus most people probably wouldn't keep track of vehicles beyond that). Essentially, you have to keep in mind which vehicles you have available to you that were police cars and use only those to complete the sub-mission, or find a police car to enter that will turn into a random vehicle (therefore maintaining the spirit of the original mission as much as possible while still giving you random cars).

Can still change later if it ends up being confusing / disliked.