General: Add Trap.asi
* Trap.asi will display a message box stating that the user has
downloaded the source code of Rainbomizer and prompt them to read the
installation instructions including a link to the latest releases page
for them to download the actual asi from. This should hopefully reduce
the number of complaints that arise from people downloading the source
code thinking it's the real thing.